Menambahkan halaman trend nilai, TODO : Publikasi

+++ lib/chart/developer_series.dart
@@ -2,7 +2,7 @@ import 'package:charts_flutter/flutter.dart' as charts;
 
 class DeveloperSeries {
   final String year;
-  final int developers;
+  final double developers;
   final charts.Color barColor;
 
   DeveloperSeries(
